% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxscore_data.R
\docType{data}
\name{boxscore}
\alias{boxscore}
\title{Simulated NBA box scores}
\format{A data.frame with 50 observations on 12 variables.
  \itemize{
    \item{code{AQ1}}{Home Team A score at the end of the 1st quarter [numeric]}
    \item{code{AQ2}}{Home Team A score at the end of the 2nd quarter [numeric]}
    \item{code{AQ3}}{Home Team A score at the end of the 3rd quarter [numeric]}
    \item{code{AQ4}}{Home Team A score at the end of the 4th quarter [numeric]}
    \item{code{BQ1}}{Opponent score at the end of the 1st quarter [numeric]}
    \item{code{BQ2}}{Opponent score at the end of the 2nd quarter [numeric]}
    \item{code{BQ3}}{Opponent score at the end of the 3rd quarter [numeric]}
    \item{code{BQ4}}{Opponent score at the end of the 4th quarter [numeric]}
    \item{code{Diff1}}{Score differential at the end of the 1st quarter [numeric]}
    \item{code{Diff2}}{Score differential at the end of the 2nd quarter [numeric]}
    \item{code{Diff3}}{Score differential at the end of the 3rd quarter [numeric]}
    \item{code{Diff4}}{Score differential at the end of the 4th quarter [numeric]}
  }}
\source{
simulation
}
\usage{
data(boxscore)
}
\description{
Simulated data on 50 NBA home games for Team A.  Scores for Team A
and their opponent, as well as the differences between them, are
recorded after each of 4 quarters.
}
\keyword{datasets}
