% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_summary.R
\name{beta_binomial_summary}
\alias{beta_binomial_summary}
\title{Summarize features of a Beta-Binomial model}
\usage{
beta_binomial_summary(a, b, x = NULL, n = NULL, posterior = FALSE)
}
\arguments{
\item{a, b}{Shape parameters of the Beta prior (Beta(a,b))}

\item{x, n}{Observed number of successes & trials}

\item{posterior}{logical; if TRUE, the posterior pdf is summarized}
}
\description{
Summarize key features of the prior and posterior pdfs for a Beta-Binomial model.
}
\examples{
library(ggplot2)
beta_binomial_summary(a=1, b=1)
beta_binomial_summary(a=1, b=1, x=5, n=10, posterior=TRUE)
}
